finished parser 100% , TODO: leaks and norminette
