+++ Makefile
@@ -1,5 +1,5 @@
 NAME = cub3d
-SRC = parser.c arr.c
+SRC = main.c arr.c input_data.c parser.c get_identifier.c get_map.c validate_map.c utils.c
 # SRC = cub3d.c raycast.c parser.c
 
 CFLAGS      = -Wall -Wextra -Werror
