+++ map.cub
@@ -14,12 +14,12 @@ C 7,8,8
         1011000001110000000000001
         1001000000000000000000001
 111111111011000001110000000000001
-100000000011000001110111111111111
+100000000000000001111111111111111
 11110111111111011100000010001
-11110111111111011101010010001
+11110111111111011101010000001
 11000000110101011100000010001
-10000000000000001100000010001
-10000000000000001101010010001
+10000000000000001000000010001
+10000000000000000001010010001
 11000001110101011111011110N0111
-11110111 1110101 101111010001 
+11110111 1110101 100000000001 
 11111111 1111111 111111111111
